Ticket #2519. Implementation of Device, Geo, User, Data, Segment, Regs objects. Fix case class code style, move builder classes outside of objects, replace "builder" method with "apply"

+++ src/main/scala/com/bitworks/rtb/model/request/Geo.scala
@@ -26,8 +26,14 @@ import com.bitworks.rtb.model.constant.LocationType
   * @param ext Placeholder for exchange-specific extensions to OpenRTB.
   */
 case class Geo(
-                lat: Option[Double], lon: Option[Double], `type`: Option[LocationType.Value],
-                country: Option[String], region: Option[String], regionfips104: Option[String],
-                metro: Option[String], city: Option[String], zip: Option[String],
-                utcoffset: Option[Int], ext: Option[Any]
-              )
+    lat: Option[Float],
+    lon: Option[Float],
+    `type`: Option[LocationType.Value],
+    country: Option[String],
+    region: Option[String],
+    regionfips104: Option[String],
+    metro: Option[String],
+    city: Option[String],
+    zip: Option[String],
+    utcoffset: Option[Int],
+    ext: Option[Any])
